requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.3.0
certifi==2025.6.15
frozenlist==1.7.0
greenlet==3.2.3
idna==3.10
magic-filter==1.0.12
multidict==6.6.2
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
SQLAlchemy==2.0.41
typing-inspection==0.4.1
typing_extensions==4.14.0
yarl==1.20.1
